fix:  remove parameters from Sequelize.INTERGE (2) seed and model Plans; fixed error:   PostgresSQL does not support INTEGER with LENGTH

+++ API/src/app/models/Plans.js
@@ -5,7 +5,7 @@ class Plans extends Model {
     super.init(
       {
         title: Sequelize.STRING,
-        duration: Sequelize.INTEGER(2),
+        duration: Sequelize.INTEGER,
         price: Sequelize.DOUBLE,
       },
       {
